+++ src/cheat/game/SpriteManager.h
@@ -32,6 +32,7 @@
 
 extern VarBoolean hud_render;
 extern VarBoolean hud_render_current_weapon;
+extern VarBoolean hud_render_velocity;
 
 extern VarBoolean hud_color_enable;
 extern VarColor hud_color;
@@ -73,7 +74,7 @@ public:
 	bool is_valid() const { return !m_screen_dimensions.IsZero() && !m_screen_pos.IsZero() && m_atlas_accruing; }
 
 	// Rendering stuff
-	void render_additive(const Vector2D& position, const CColor& color, const uint32_t frame = 0) const;
+	void render_additive(const Vector2D& position, const CColor& color, bool dim, const uint32_t frame = 0) const;
 
 	std::string m_name;
 
@@ -123,24 +124,32 @@ public:
 
 	void handle_color_change(hl::HSPRITE_t spriteId, int& r, int& g, int& b);
 
+	// expects 0-255 range
+	static void scale_hud_colors(CColor& color, bool dim);
+
 private:
 	void initialize();
 
 	bool initialize_engine_sprite_data();
 
-	// Rendering
+	//
+	// rendering
+	//
+
 	void render_current_weapon_sprite() const;
+	void render_velocity() const;
+
 	Vector2D render_digit(uint8_t digit, const Vector2D& position, const CColor& color) const;
-	Vector2D render_number(int64_t number, const Vector2D& position, const CColor& color) const;
+	Vector2D render_number(int64_t number, const Vector2D& position, const CColor& color, bool centerize = false) const;
 	template<uint8_t Precision> requires(Precision < std::numeric_limits<double>::max_digits10 && Precision > 0)
-	Vector2D render_float_number(double number, const Vector2D& position, const CColor& color) const;
+	Vector2D render_float_number(double number, const Vector2D& position, const CColor& color, bool centerize = false) const;
 
 private:
 	uint32_t m_resolution_scale;
 
 	static std::unordered_map<std::string, SpriteAtlas_t> s_sprite_atlas_map;
 	static std::unordered_map<std::string, Sprite_t> s_sprite_map;
-	constexpr static inline Vector2D m_FontSize = { 20, 25 };
+	Vector2D m_cstrike_sprite_font_size;
 };
 
 #endif // SPRITEMANAGER_H